♻️ refactor : DTO 아우터 인터페이스 네이밍변경 (#5)
- BookRequest 내부 이너클래스로서 Action 을 설명하는 것으로 변경한다
