+++ src/main/java/com/devcourse/checkmoi/domain/book/dto/BookRequest.java
@@ -1,10 +1,10 @@
 package com.devcourse.checkmoi.domain.book.dto;
 
-import com.devcourse.checkmoi.domain.book.dto.UpdateBookRequest.CreateBook;
+import com.devcourse.checkmoi.domain.book.dto.BookRequest.CreateBook;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 
-public sealed interface UpdateBookRequest permits CreateBook {
+public sealed interface BookRequest permits CreateBook {
 
     // TODO : api 로 받아온 데이터들에 대해서는 어떤 검증조건이 있어야 할지 아직 미정상태
     record CreateBook(
@@ -16,7 +16,7 @@ public sealed interface UpdateBookRequest permits CreateBook {
         String pubDate,
         String isbn,
         String description
-    ) implements UpdateBookRequest {
+    ) implements BookRequest {
 
         @Builder
         public CreateBook {
